rename precise_float
