+++ source/core/plugin/PreciseFloat.h
