+++ source/gui/UnitConverter.h
@@ -3,7 +3,7 @@
 #include <vector>
 #include <map>
 #include <string>
-#include <core/plugin/precise_float.h>
+#include <core/plugin/PreciseFloat.h>
 
 class UnitConverter
 {
